Update styles.css for issue1
PROJECT FOLDER - bug-beatles/modern_chair
CURRENT STATE - Currently, when we select a color for the chair. it does not match the color of the chair displayed on the screen afterwards.
DESIRED STATE - Fix the above explained issue so that the displayed color of the chair matches the selected color option when making a choice from the available color options for the chair.

I made changes from line 90 onwards. I synchronised the images with the labels.